cloudrun/requirements.txt
streamlit
gclient-service-account-auth
google-cloud-storage
google-cloud-bigquery